Implement Structs, default init
Implement assign and get from struct and fields
Handle default initialization in variabel

Struct can have NIL field (default) or be created as a "new" struct which default initializes it's fields, the value itself is nil if the struct is nil, but there is a distinction for typing sake.
That is comparison between variables of different structs that are both nil cannot be done. Because expressions expand all the way to the value, type checking the value itself (the struct wrapper around NIL) is helpful. The variable type is expanded first so it's less feasible to do type-checking based on variable types when it comes to expression evaluation.

Change the type checking to simply check if the struct types match instead of checking field matchings