Fix: NIL wrapped structs of same type are equal even if object references are different
This might be an artificat of the way I'm creating default structs.

I wonder if the canonical implementation has a NIL struct of that type that every newly declared struct points to the same NIL struct. In my case I create a new NIL reference for each but this is admittedly a worse approach, but it was convenient.