Add exception handling
Add a special Exception type called BrewinException to specifically catch errors raised by Brewin code and not the interpreter itself

It stores the Brewin exception_type which can be accessed to compare against catchers.

Wrap the main function in a try catch that acts as the final catch if there are no matching catchers in the actual code which then throws the FaultError appropriately.