Fix Lazy Evaluation
My previous lazy evaluation implementation was incorrect because it only considered lazy expressions stored in variables.

However, it is possible for arguments and return values to be lazily evaluated as well.

I have now updated the implementation to pass arguments as lazy expressions and to return lazy expressions.

A problem arises with my previous implementation when it comes to return values:

For return values, because they may be accesssed without retrieving a variable value with the VAR_NODE, I can't rely on a single Element node to trigger the lazy evaluation.
This would be different if I don't use primtive values throughout and instead wrapped my values in a Value class which could then be used to trigger the lazy evaluation whenever a value is accessed.

The main problem is that at the time of returning from a function it is not determiniable if the return value will be used or not, unlike a VAR_NODE which is immediately requesting the value of a variable, thus triggering the lazy evaluation.

Only the caller of the function knows if the return value will be used or not, so the lazy evaluation should be triggered by the caller once the value is used and so after I call an evaluate expression. In case it is an fcall that returned a lazy expression, and if I check that it is, then I call the evaluate on the LazyExpression which may be a bunch of nested LazyExpressions in the case of a function returning a call to another function, etc. So I use a while loop, too fully evaluate the lazy expression.

This feels not very clean, because essentially in every place where I call evaluate_expression, I also try to get_value() to evaluate a potential lazy expression.

When I snapshot the scope, I no longer deep copy, this because a lazy expression that was evaluated as a result of another lazy expression should update itself to be evaluated and cache it's value. In order to do this, the nested lazy expression should not be a deep copy and instead should be a reference such that if that lazy expression were to be evaluated elsewhere it would update the original lazy expression. But the dict for that scope should be copied so that the original scope is not modified when it comes to assignment. Since all other values are primitives there is no concern about mutating them so they are shallow copied.